feat(select): Finish heapsort
This modifies the compare to always return the amount
of comparisons, regardless if there is an error. This was
neccessary to analyze the selection and sort algorithm
independantly. Also, error reporting was turned off for
the sake of a clean terminal session. I might want to turn
this back on if I continue developing.